requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
sentence-transformers==3.3.1
faiss-cpu>=1.9.0
streamlit==1.40.1
python-dotenv==1.0.1
pypdf==5.1.0
langchain-huggingface==0.1.2
